NEW: Bringing up a sample Flask Poll app + unpublished Flask app.
Since I didn't have any good demo in Flask (mine was a simple app with no
storage), I dediced to get a sample poll app on github. Specifically this one:

https://github.com/kalise/flask-vote-app.git

For that particular app, I had to remove the session cookie because it was
not working properly after the app was being shut down and didn't show cleary
to the user that he already voted. Who knows in the future I could solve it and
give the commit back to the original author?

This app lacks testing, so I have to pick up my own test samples and now I have
to fix the tests before working on Vagrant.